Fix crash from concurrent nb::make_iterator<> under free-threading.
The `State` class used by make_iterator is constructed lazily, but
without locking it is possible for the caller to crash when the class
type is only partially constructed. This PR adds an ft_mutex around the
binding of the State type.

My initial instinct was to use an nb_object_guard on `scope`.
Unfortunately this doesn't work; I suspect PyEval_SaveThread() is called
during class binding and that releases the outer critical section.
