+++ tests/test_make_iterator.py
@@ -1,5 +1,5 @@
 import test_make_iterator_ext as t
-import pytest
+from common import parallelize
 
 data = [
     {},
@@ -28,6 +28,10 @@ def test03_items_iterator():
         m = t.StringMap(d)
         assert sorted(list(m.items())) == sorted(list(d.items()))
         assert sorted(list(m.items_l())) == sorted(list(d.items()))
+
+
+def test03_items_iterator_parallel(n_threads=8):
+    parallelize(test03_items_iterator, n_threads=n_threads)
 
 
 def test04_passthrough_iterator():
