+++ tests/test_thread.py
@@ -1,29 +1,6 @@
 import test_thread_ext as t
 from test_thread_ext import Counter
-
-import threading
-
-# Helper function to parallelize execution of a function. We intentionally
-# don't use the Python threads pools here to have threads shut down / start
-# between test cases.
-def parallelize(func, n_threads):
-    barrier = threading.Barrier(n_threads)
-    result = [None]*n_threads
-
-    def wrapper(i):
-        barrier.wait()
-        result[i] = func()
-
-    workers = []
-    for i in range(n_threads):
-        t = threading.Thread(target=wrapper, args=(i,))
-        t.start()
-        workers.append(t)
-
-    for worker in workers:
-        worker.join()
-    return result
-
+from common import parallelize
 
 def test01_object_creation(n_threads=8):
     # This test hammers 'inst_c2p' from multiple threads, and
